Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Could this be a type of deployment so we maintain both functionality for now?

Is your comment similar to Jacques? If it is a generic comment, may be you are not aware of the background for the change. For deployment, when a pod exits (crash, deleted, whatever), all changes to the disk are lost - which in the case of executors means that the reflections on it are also lost. So, one of the solution was to make the executors stateful set so that even when it crashes, have the disk persist and associated with the same host.

Patch-set: 1