Update patch set 1
Patch Set 1:

> Re adding to README - do we need to? Then we need to describe every
 > property in values.yaml in the README. In the values.yaml, there is
 > a description and a link to k8s doc for more info.

I feel this is a feature that will cause some confusion for customers and generate support tickets that can be avoided with more guidance.  Having a simple example in the README should help a lot.  Just an abbreviated version of the linked k8s doc would suffice.

Patch-set: 1