DX-16610: Remove the strategy from templates
Statefulsets attribute is updateStrategy and the default for that
is RollingUpdate. So, deleting those lines would be sufficient
and not loose any functionality.

Kubernetes client libraries to v1.14 seems to be more strict than
earlier version and throws an error with those lines in there.

Dropping the lines works with earlier versions of helm.

Change-Id: I727adc50a883a4801e737aec0be6b84788f6f357