Update patch set 1
Patch Set 1: Code-Review-2

We don't want our customers to ever deploy a single C+E node. We shouldn't provide a public option to do this because it will signal that we support this sort of deployment.

If we want to support it internally for specific pre-qualified use cases, that's something that we should do independently of this chart. Also -- note that there is a new version of the chart that I've put up for review here: https://gerrit.drem.io/gerrit/c/dremio-cloud-tools/+/19086

Patch-set: 1
Reviewer: Gerrit User 1000118 <1000118@8ca9cf2d-83f4-421d-8552-b009c9e056de>
Label: Code-Review=-2