Setting `imageTag` to a number causes `helm install` to fail
If you set `imageTag` to something that can be interpreted as a number (e.g. 20 or 20.1 ) then helm interprets it as a number, and the comparison on line 18 in dremio.conf  (` {{- if eq $.Values.imageTag "latest" }}`) fails because it is comparing a number to a string.

Enclosing the number in quotes coerces it into a string and the comparison is fine.